+++ tests/testit/test-formats.R
@@ -77,7 +77,7 @@ test_format("rsos")
 test_format("rss")
 test_format("sage")
 test_format("sim")
-test_format("springer")
+test_format("springer", skip = !rmarkdown::pandoc_available("2.11.4"))
 test_format("tf")
 test_format("trb")
 # Deactivate because of https://github.com/rstudio/rticles/issues/523